monitoring_system/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
cryptography==41.0.7
gunicorn==21.2.0
python-decouple==3.8
django-filter==23.3
channels==4.0.0